Fix possible crash with low value of file descriptors

+++ include/siri/grammar/gramp.h
@@ -26,8 +26,8 @@
 
 
 #if CLERI_VERSION_MINOR >= 12
-#define CLERI_NODE_DATA(__node) ((intptr_t)(__node)->data)
-#define CLERI_NODE_DATA_ADDR(__node) ((intptr_t *) &(__node)->data)
+#define CLERI_NODE_DATA(__node) ((int64_t)(__node)->data)
+#define CLERI_NODE_DATA_ADDR(__node) ((int64_t *) &(__node)->data)
 #else
 #define CLERI_NODE_DATA(__node) (__node)->result
 #define CLERI_NODE_DATA_ADDR(__node) &(__node)->result
